pyproject.toml
[project]
name = "context-engineering-labs"
version = "0.1.0"
description = "OpenAI-first context engineering labs: budgeting, packing, observability, red-teaming."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.0.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "rich>=13.7.1",
  "black>=25.12.0",
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "pytest-xdist>=3.6.1",
  "ruff>=0.5.0",
]

[tool.pytest.ini_options]
addopts = "-q"
markers = [
  "requires_llm: tests that require a real OpenAI API key",
]

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/context_engineering"]
